Move components to ui folder
